test: fold Google GenAI stream failure suite
