--- src/test-kernel/agent/modelHandlers/GoogleGenAIStreamRetryEligibility.vitest.ts
@@ -1,105 +1,0 @@
-// Third-party imports
-import { describe, expect, it } from 'vitest';
-import { createPartFromText, type Content } from '@google/genai';
-import { ModelProvider } from 'llm-zoo';
-
-// Local imports
-import { noopTrace } from '@agent/trace';
-import { ModelHandlerGoogleGenAI } from '@agent/modelHandlers/google/modelHandlerGoogleGenAI';
-import { detectPartialText } from '@common/errors/sdkErrorUtils';
-import { buildTestModelConfig } from '@test/support/modelConfigTestUtils';
-
-class StreamingGoogleHandler extends ModelHandlerGoogleGenAI {
-  override getStreamingConfig(): boolean {
-    return true;
-  }
-}
-
-function createStreamingHandler(): ModelHandlerGoogleGenAI {
-  const handler = new StreamingGoogleHandler(
-    buildTestModelConfig({
-      name: 'test-google-model',
-      label: 'Test Google Model',
-      fullName: 'google/test',
-      shortName: 'google/test',
-      provider: ModelProvider.GOOGLE,
-      contextWindow: 4096,
-      capabilities: { supportsTokenCounting: false },
-    }),
-  );
-  handler.setLogger({
-    ...noopTrace,
-    openStream: () => ({
-      id: 'stream-1',
-      append: () => undefined,
-      finalize: (text?: string) => text ?? '',
-    }),
-  });
-  handler.setOutputStreaming(true);
-  return handler;
-}
-
-/**
- * A stream that yields one chunk of visible text, then throws mid-stream —
- * simulating a dropped connection *after* content has already been produced.
- */
-function createFakeClientWithMidStreamFailure(
-  visibleText: string,
-  failure: Error,
-): any {
-  return {
-    chats: {
-      create: () => ({
-        sendMessageStream: async () =>
-          (async function* () {
-            const chunk: any = {
-              candidates: [
-                {
-                  content: {
-                    role: 'model',
-                    parts: [createPartFromText(visibleText)],
-                  },
-                },
-              ],
-            };
-            yield chunk;
-            throw failure;
-          })(),
-        sendMessage: async () => {
-          throw new Error(
-            'sendMessage should not be called when streaming is enabled',
-          );
-        },
-      }),
-    },
-    models: {},
-  };
-}
-
-describe('ModelHandlerGoogleGenAI streaming failure diagnostics', () => {
-  it('retains partial text from a mid-stream failure', async () => {
-    const handler = createStreamingHandler();
-    const failure = new Error('relay connection dropped mid-stream');
-
-    const messages: Content[] = [
-      { role: 'user', parts: [createPartFromText('Hi there')] },
-    ];
-
-    let caught: unknown;
-    try {
-      await handler.createResponse({
-        client: createFakeClientWithMidStreamFailure(
-          'partial text before the drop',
-          failure,
-        ),
-        messages,
-        temperature: 0,
-      });
-    } catch (err) {
-      caught = err;
-    }
-
-    expect(caught).toBe(failure);
-    expect(detectPartialText(failure)).toBe('partial text before the drop');
-  });
-});
